Fix terrible bug with handling of the ownership frames
I was incorrectly relying on the destructor of the ownership frame
to sanity check it's presence in the ownership stack. This is of course
logically, a circular dependency and was never being satisfied
previously, failing the deallocation assertion.

Fixed by adding an ownership lock which which triggers the popping
of the ownership frame from the ownership stack.

Getting rid of the shared_ptr'ness of the ownership frame is not
possible, because we would have needed the address of
the allocated space before it was allocated. With the shared_ptr, we can
go back and update the address and it will automagically be reflected to
everyone holding a shared_ptr to the same object. Wouldn't have been
possible by neither a unique_ptr nor a bare bones uintptr_t value of
the pointer.